controller/requirements.txt
contourpy==1.3.3
pyyaml
cycler==0.12.1
fonttools==4.62.1
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.3
packaging==26.0
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pyzmq==27.1.0
scipy==1.17.1
six==1.17.0
